Revisit media selection implementation
